requirements.txt
# Core
torch>=2.2
torchvision>=0.17
numpy>=1.25
scipy>=1.11
scikit-learn>=1.3
Pillow>=10.0

# Metrics & plots
scikit-image>=0.21
matplotlib


# utils
tqdm>=4.66
python-dotenv>=1.0.1
pyyaml>=6.0